refactor: StudyScreen 컴포저블 분리
